fix position numbering, add tests
In #4, @steletvinicius pointed out that the position numbers were wrong
for HLA-A prot.

Indeed, he was right. So, I went back to the parsing code and rewrote it
again.

This time, I added tests to make sure that the position numbers are
correct:

    new file:   tests/testthat/test-position_numbers-prot.R

I also started a new file for the gen sequences:

    new file:   tests/testthat/test-position_numbers-gen.R